Allow comments (lines starting with # character) and blank lines in openmw.cfg.  Should resolve Feature #2535.
- controlled via a checkbox in launcher settings

+++ apps/launcher/maindialog.cpp
@@ -490,7 +490,16 @@ bool Launcher::MainDialog::writeSettings()
     // Game settings
     QFile file(userPath + QString("openmw.cfg"));
 
-    if (!file.open(QIODevice::ReadWrite | QIODevice::Text | QIODevice::Truncate)) {
+    QIODevice::OpenMode mode(0);
+    bool keepComments = mLauncherSettings.value(QString("Settings/keep-comments"), QString("true"))
+                        == QLatin1String("true");
+
+    if (keepComments)
+        mode = QIODevice::ReadWrite | QIODevice::Text;
+    else
+        mode = QIODevice::ReadWrite | QIODevice::Text | QIODevice::Truncate;
+
+    if (!file.open(mode)) {
         // File cannot be opened or created
         QMessageBox msgBox;
         msgBox.setWindowTitle(tr("Error writing OpenMW configuration file"));
@@ -504,9 +513,15 @@ bool Launcher::MainDialog::writeSettings()
     }
 
     QTextStream stream(&file);
-    stream.setCodec(QTextCodec::codecForName("UTF-8"));
 
-    mGameSettings.writeFile(stream);
+    if (keepComments)
+        mGameSettings.writeFileWithComments(file);
+    else
+    {
+        stream.setCodec(QTextCodec::codecForName("UTF-8"));
+        mGameSettings.writeFile(stream);
+    }
+
     file.close();
 
     // Graphics settings
